double precision

+++ src/canari/data_struct.py
@@ -41,8 +41,8 @@ class LstmOutputHistory:
         Args:
             look_back_len (int): Number of time steps to keep in history.
         """
-        self.mu = np.zeros(look_back_len, dtype=np.float32)
-        self.var = np.ones(look_back_len, dtype=np.float32)
+        self.mu = np.zeros(look_back_len)
+        self.var = np.ones(look_back_len)
 
     def update(self, mu_lstm, var_lstm):
         """
